// src/shared/api/utils/apiRoutes.ts
enum ApiRoutes {
  attack = "/attack",
  attackCreate = "/attack/create",
  attackStatus = "/attack/status",
  groupvuln = "/gr3pr",
  certvuln = "/cert",
  dchost = "/dc",
  filevuln = "/filev",
  agents = '/agents',
  kerberos = "/kbr",
  load = "/load",
  scanvuln = "/scanv",
  users = "/users",
  logAttack = '/analytics/log',
  anyltUsers = '/analytics/users',
  statistics = "/statistics",
  check = "/vuln",
  accounts= '/accounts',
  hosts= '/analytics/hosts',
  dbDefects='/database/defects',
  dbActions='/database/actions'
}

export default ApiRoutes;